production/nlx-model-audio-gender/requirements.txt
kafka-python
numpy
matplotlib
scipy
sklearn
hmmlearn
simplejson
eyed3
pydub
nltk
colorama
libmagic
librosa
pymongo
dnspython
requests
